pyproject.toml
[tool.poetry]
name = "zsa-keymapp-python"
version = "0.1.0"
description = "Control your ZSA keyboard using the Keymapp API (Protobuf) in Python"
authors = ["Jordan Munch O'Hare <github@jordan.munchohare.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.15.1"
mypy-protobuf = "^3.6.0"
grpcio = "^1.69.0"
grpcio-tools = "^1.69.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
